Develop

+++ internal/service/db/data_service.go
@@ -28,8 +28,8 @@ func NewRecordService() {
 func (s *DataService) ProcessCSVFiles(records_file_names []string) error {
 	records := s.record_service.GetRecords(records_file_names)
 	s.vehicle_service.SaveAll(records)
-	equipment_ids := Map(records, func(record csv.Record) int32 {
-		return int32(record.Equipment)
+	equipment_ids := Map(records, func(record csv.Record) int64 {
+		return record.Equipment
 	})
 	s.equipment_service.SaveAll(&equipment_ids)
 	return nil
